requirements.txt

python-dotenv
praw
google-generativeai
Jinja2